Passwords, as user input, may not be 7-bit ASCII -- so we need 	to treat them as unsigned char*'s.  No surprises there; this 	just regularises the usage so we don't get compilation messages.
git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/1.3.x@83037 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/http_core.c
@@ -3101,7 +3101,7 @@ static int default_handler(request_rec *r)
 	    AP_MD5_CTX context;
 	    
 	    ap_MD5Init(&context);
-	    ap_MD5Update(&context, (void *)mm, r->finfo.st_size);
+	    ap_MD5Update(&context, (void *)mm, (unsigned int)r->finfo.st_size);
 	    ap_table_setn(r->headers_out, "Content-MD5",
 			  ap_md5contextTo64(r->pool, &context));
 	}
